# gen/blog.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "blog.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/posts": {
      "post": {
        "operationId": "Blog_GetPosts",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/blogPost"
                },
                "error": {
                  "$ref": "#/definitions/runtimeStreamError"
                }
              },
              "title": "Stream result of blogPost"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/blogPage"
            }
          }
        ],
        "tags": [
          "Blog"
        ]
      },
      "put": {
        "operationId": "Blog_CreatePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/blogCreatePostResult"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/blogPost"
            }
          }
        ],
        "tags": [
          "Blog"
        ]
      },
      "patch": {
        "operationId": "Blog_EditPost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/blogEditPostResult"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/blogPost"
            }
          }
        ],
        "tags": [
          "Blog"
        ]
      }
    },
    "/v1/posts/{value}": {
      "delete": {
        "operationId": "Blog_DeletePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/blogDeletePostResult"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Blog"
        ]
      }
    },
    "/v1/posts/{value}/like": {
      "delete": {
        "operationId": "Blog_UnlikePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/blogUnlikePostResult"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Blog"
        ]
      },
      "put": {
        "operationId": "Blog_LikePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/blogLikePostResult"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Blog"
        ]
      }
    }
  },
  "definitions": {
    "blogCreatePostResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "created_id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "blogDeletePostResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "blogEditPostResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "blogLikePostResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "blogPage": {
      "type": "object",
      "properties": {
        "offset": {
          "type": "string",
          "format": "uint64"
        },
        "limit": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "blogPost": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/blogPostId"
        },
        "author": {
          "$ref": "#/definitions/blogUser"
        },
        "body": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "uint64",
          "description": "Seconds since 1 Jan, 1970 00:00 UTC (unix time)\nConverted to client date and time with timezone\nat front end."
        },
        "num_liked": {
          "type": "integer",
          "format": "int64"
        },
        "liked": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/blogUser"
          }
        }
      }
    },
    "blogPostId": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "blogUnlikePostResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "blogUrl": {
      "type": "object",
      "properties": {
        "href": {
          "type": "string"
        }
      }
    },
    "blogUser": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/blogUserId"
        },
        "nickname": {
          "type": "string"
        },
        "photo_url": {
          "$ref": "#/definitions/blogUrl"
        }
      }
    },
    "blogUserId": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "runtimeStreamError": {
      "type": "object",
      "properties": {
        "grpc_code": {
          "type": "integer",
          "format": "int32"
        },
        "http_code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "http_status": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}